Remove UnsupportedBlock in favor of core/missing
